Social network more customisable
Instead of filling a preconstructed string with links to specific values
in the config file, the /social command look for each social link in the
dictionnary. This way, it is possible to remove, add, and fully customise
those links.
As dictionnaries don't keep the initial order of entries from the config
file, they are now sorted by alphabetical order.